Add nice names for known resource servers
Each resource server listed in our scope requirements map has a
nice_server_name attribute associated with it, resulting in the
expected mapping

  auth -> Globus Auth
  transfer -> Globus Transfer

but notably also

  $timer_service_id -> Globus Timers

This enables better output when a command fails the `requires_login`
check, as the user is prompted with a known short name rather than a
less readable resource_server value.

To implement this, the MissingLoginError class now depends upon the
scope requirements. But this could be refactored in the future to make
the integration between the two more loosely coupled.

The names are sorted to provide good guarantees to the testsuite.