Control server-timing display with hidden option
`--show-server-timing` is a hidden option which will enable the
rendering of Server-Timing headers. This replaces the env var which
was used in the first draft.

To support this, a new bool field gets attached to the state object
and a new hidden option goes into the common opts to flip that bool
flag. A helper in the "context" module for 'globus_cli.termio' can
inspect the state object for this flag.

Also, fixup some type annotations and comments which could be
trivially improved.