Refactor ServerTiming parse and add initial tests
A framework is now in place to support multiple versions of the
Server-Timing header spec, although the only one which is supported is
still the 2017 version in which there is no `dur` key for the
duration, and the format is primarily `name=duration`.

Future parsers could be added in the future and wrapped such that they
are equally well supported. At that point, it would be possible to try
multiple parses to support multiple different spec versions if
desired.