Add JSONStringOrFileV2 to improve JSON parsing
This is a new param type which is intended to replace JSONStringOrFile
at all usage sites. As a first step in the process, introduce the new
type, authored to be a fresh implementation with very similar
behavior, and copy the existing JSONStringOrFile tests to produce a
compatibility test which verifies that the new object behaves
similarly.

As a key difference, the new implementation does not require the
`file:` prefix, but will treat it as a disambiguator which means "do
not try a JSON parse of this data".

The order of precedence for the parse is:
- `file:` prefix (if present)
- file loading
- JSON parse (if no such file)

Errors are treated somewhat carefully, to try to ensure that the JSON
parse will not hide a missing file error and misreport it. However,
this requires ultimately attempting to detect whether or not the input
was intended to be JSON or a filename. This is done with an
intentionally simplistic heuristic: check if the first char is `{`.

As an improvement to the interface, `JSONStringOrFileV2` does not
return the raw loaded JSON data, but places it into a dataclass
wrapper which pairs the loaded data with the filename. Although many
usages will not benefit from this explicit treatment, it makes it
possible to give back an error on malformed data which refers to the
file containing said data.

Additionally, JSONStringOrFileV2 handles `-` to indicate stdin,
opening up additional possible usages.