Update user-credential from-json parsing
Update these to use JSONStringOrFileV2. Additionally, now that the
argumetns are being annotated, add these commands to the click
annotation checker.

Now that the data type is known on these data to be any valid JSON,
mypy is able to detect and flag non-dict data as a possibility. To
resolve this, add new runtime checks which confirm that the data is
object-shaped.
New tests exercise the runtime checks.