Convert 'flows start' to JSONStringOrFileV2
Update the parsing to use JSONStringOrFileV2, update the docstring to
show a filename without the `file:` prefix, and update the command
internals slightly to handle the new parsed data object.

This catches a potential type error in which non-dict data could have
been parsed and would have been passed through. Now that the types are
more precise, it is flagged by mypy as a type checker error. To
resolve, a new check is added to confirm that the input data is
object-shaped.
A new test exercises this check.