Use JSONStringOrFile in search commands
This replaces the use of `click.File('r')` combined with explicit
calls to `json.load`. The end result is that we are using the
`JSONStringOrFile` helper in all locations where it is valid to apply,
and the datatypes within the relevant search commands
(delete-by-query, query, and ingest) are knowable for mypy.

As a result of these improvements, we need some additional changes.
Runtime checks for non-dict data (as in other uses of ParsedJSONData),
and in one case a widening type cast for `list[str]` to `list[JSON]`.