Improve CLI reflection tools
- Move reflection tooling to a dedicated module, `globus_cli.reflect`,
  pulling from the doc generator and `globus_cli.utils`
- Use `reflect` to improve the implementation of the click annotation
  checker test to automatically pick up on any new modules

Once this is done, road-test the changes to the annotation checker by
removing modules from the deny-list to ensure that the checker picks
them up.
Using `globus endpoint activate`, verify that the checker can still be
used to identify and fix a minor issue after the refactor.

This also enables the checker on many modules where it already passes,
expanding the list of checks.