Improve error handling for a malformed `--batch`
Batch input parsing via the shlex subcommand parsing method did not
create a new click context and allowed the default
`ClickException.show()` call in command execution to run. As a result,
errors were rendered in a confusing way, showing 'globus' as the
command name and showing the failed parse as belonging to it. This
confuses the fact that this is a parse from an internal subcommand
invocation.

To resolve, the following changes are made:
- shlex_process_stream now requires a new argument, the "name" to use
  for the parse in progress in case of failure
- Invocations are done by manually constructing the subcommand context
  and calling `invoke()`, rather than via `main()`. This gives us
  better control over failures
- Errors are explicitly caught, emitted in a custom format, and then
  converted to an `Exit(2)`

Tests for the above exercise the happy-path and sad-path cases and
verify that the behavior fo shlex_process_stream itself is as
expected.

Resolves #786