Allow `--public/--private` on GCP
The `--public` and `--private` flags used to be restricted to GCSv4
Host Endpoints only. This restriction is no longer in sync with the
service requirements.

Remove the restriction from the validation code in
`globus_cli.commands.endpoint._common`, which is used for both
`endpoint update` and `endpoint create` validation.
Test updates remove tests which look for these usages to be rejected
and add `--private` to the tested parameters for GCP create/update
commands.