Update annotations for 'endpoint server' commands
This improves our annotation coverage to include this full category of
commands.

For the most part, the changes are new annotations without runtime
changes, but a few modifications improved and simplified things.
The largest change is a restructuring of the `add` and `update` shared
options, which are slightly rewritten to make the annotations simpler.

The custom TransferClient helper for `get_endpoint_w_server_list` has
its annotation updated to be slightly more precise.