Improve type annotations for 'collection' commands
Four basic changes are made here. The first two are relatively
standard:
1. Remove mypy.ini ignore rules for 'collection' commands
2. Remove click type checking exemptions for 'collection' commands

The next two are tweaks to get the custom ChoiceSlugified type used in
the collection list command to pass the click-type-checker.

3. Change precedence order on the click type checker to prefer
   AnnotatedParamType.get_type_annotation over any recognized built-in
   click types

This ensures that a type which inherits from a click type can set an
explicit annotation and have that value checked instead of the
"normal" type for click deduction.

4. Implement `get_type_annotation` for ChoiceSlugified to normalize
   the choice values

The goal of the ChoiceSlugified type is to present 'foo-bar' as the
CLI interface for passing 'foo_bar'. It normalizes `-_` to `-` to
match the internal choice, then re-norms `-` to `_` as its output.
Therefore, a command accepting this parameter type should expect
`_`-normed strings as inputs. That is exactly what the new
`get_type_annotation` method implements.

In addition to the various type-checking changes, there is a very
minor runtime change in that one command which accepted non-UUID
endpoint-id values (any string) now requires UUID values by means of
the common `endpoint_id_arg` decorator.