First draft of complex run resume
This makes it possible to resume a run which is INACTIVE due to
ConsentRequired. The current user's consents are checked against the
Consent Forest data returned by Auth in order to gate the resume
between requiring a `session consent` flow or not.

In support of this, some enhancements are made to the consent
containment check to support parsed scopes.