Replace last distutils usage
This usage made the CLI py3.12 incompatible. Remove distutils usage
for string parsing and reproduce that same utility (with
near-identical semantics) in globus_cli.utils.

The invalid-raise behavior is now baked into the getter with a
`raising=True` flag, so that the details about the env var name are
not leaked as much out of the utility.