Convert to ConnectorTable for resolution
Rather than keeping a CLI-specific dict of IDs and names, use the
ConnectorTable object provided by the SDK to resolve IDs to names.

This changeset also eliminates a never-used method for converting
names to IDs, which is functionality provided by the ConnectorTable if
we should need it.

In addition to test refactoring, the relevant test module is renamed.

The changelog notes this as a rendering fix, since Dropbox and HPSS
were previously missing from the data but are included in the
ConnectorTable. So, effectively two entries have been "added" to this
formatter.