Update to use the Flows 'all' scope
This is defined to require a new login (`min_contract_version` is bumped
to 3) in order to ensure that all future interactions with Flows have
scopes matching what is listed in the code.

It would be possible to roll out this change without bumping the
contract-version, since no new permissions are needed in Flows. However,
that would leave us in a degraded state moving forward for development,
in which we wouldn't be able to assert that the tokens used by a CLI
instance really match what's listed in our application.