Minor changelog copyedit fixes
This appears to have been written with rst styling, and sphinx-issues
styled PR links.

Because the CLI changelog is rendered to the docs site, it needs to be
formatted as asciidoc (and because scriv doesn't have native asciidoc
support, we use markdown which we later convert).

The choice as to whether or not to link PRs is an old one, and may
deserve to be revisited, but it is our standing practice not to link
them for the CLI.