Improve the helptext for 'collection delete'
This corrects and expands the helptext, after consultation with the
GCS team to confirm accuracy.

Because guest collection owners are the most likely to be confused
about their ability to delete, based on what we've seen, Collection
owner/admin permissions are called out first and separately from the
other roles which grant deletion capabilities.

Most of the content can be pulled from the docs for the delete API:
https://docs.globus.org/globus-connect-server/v5.4/api/openapi_Collections/#deleteCollection