Implement flow start notification policy parsing
`globus flows start` has a new option,
`--activity-notification-policy`, with specialized custom parsing. It
accepts either a comma-delimited list or a JSON blob (or file),
deriving from our existing JSON support.

A simple heuristic tries to determine which type of input is being
given, checking if it "looks like comma delimited names" and
dispatching between the parse paths as appropriate.
The same heuristic drives tab completion, letting the completer switch
between the two modes.

Unit tests confirm the functionality to ensure that it parses as
desired in a variety of scenarios, and emits appropriate errors.

The resulting parsed policy object is then simply passed to
`start_flow`.

Additionally, in order to provide good helptext for this option, the
overall helptext for `globus flows start` has been enhanced.