Refine the contributing doc + add a changelog doc
The contributing doc was a bit dated and had some phrasing which could
be construed as unfriendly.

It also stated well-established norms for the project (e.g., we are
careful about adding new 3rd party dependencies) which don't need to be
re-stated in a new-contributor-facing document.

The changelog process had effective no documentation around it,
which is rectified by introducing a changelog process doc in
`changelog.d/README.adoc`. The contributing doc references this.

The doc now covers how to run `tox` to invoke tests, without trying to
provide a comprehensive `tox` tutorial but while still showing some
important usages.

The introductory note, in particular, but much of the rest of the doc as
well, is inspired by the guidance offered by `attrs` [^1].

[^1]: https://github.com/python-attrs/attrs/blob/main/.github/CONTRIBUTING.md