Add unit tests for the UnixPrinter
These fully exercise all of the codepaths, but not necessarily all of
the interesting printing modalities which can be constructed out of those
codepaths.

All of the tests are focused on the external behavior except for one
which is a bit of "code coverage hacking" by exploiting an unrealistic
gap in logic around truly badly-behaved objects.