Fix tests for 'base_path' matching
Under SDKv4, the 'base_path' attribute is removed. The `globus api`
command continues to inject base paths for `groups` and `transfer`
commands as of now, for compatibility. Registered response mocks must
use the full paths 'transfer' and 'groups'.

As a result, test mocks which are missing the legacy 'base_path' values
need to be updated.