Remove the legacy 'get_submission_id' mock
An old mock for `get_submission_id` (as a YAML file) was still in use.
However, `globus_sdk.testing` provides a mock for this API call which can
be switched to trivially.

Additionally, several tests were using the mock which no longer need it
due to changes in SDK behavior to defer the request for a submission ID
until it is actually needed.