Improve 'bookmark list' failure test case
Migrate off of the old fixture file and onto new `globus_sdk.testing`
based response mocks, which are included in the test functions
themselves. As a direct consequence of this refactor, it is obvious that
there are two different failure modes encoded into the fixture (one soft
and one hard), but only one of the two is tested. Split the test into the
two relevant cases.

One of the new cases overlaps with the success test case tested in the
same file. This is noted in a TODO comment to remove the test redundancy.