Add a new local flake8 rule for repeated fieldnames
We can detect repeated fieldnames statically when field lists are fully
static. Some minimal assumptions are necessary, primarily that a field
list is a list of calls to `Field()` and that `Field()` means what we
think it does and is always imported under a FromImport.

This fails on the same case which is caught by runtime testing, namely
endpoint show.